Handle numbers in targetvalue for synthetics assertions
For some specific assertions operators, the target value needs to be a
number for the API to dispatch the request properly. This adds a cast
when we send data and when storing it so that those assertions can be
handled properly.

This also fixes error reporting on test update in general.

Closes #765