Use code formatting in description for attribute remapper
From "log attribute or tag", it's not clear what the permitted values of `source_type` and `target_type` are, exactly (`log attribute` and `tag`? `log-attribute` and `tag`? `attribute` and `tag`? ...). The API docs use code formatting for `attribute` and `tag` and make it clear (see option 6 in [the "create a pipeline" section][1] of the API docs). Doing the same here would increase clarity.

[1]: https://docs.datadoghq.com/api/latest/logs-pipelines/#create-a-pipeline