[datadog_service_definition_yaml] Prevent recreating resources on casing-only differences
In-depth: https://github.com/DataDog/terraform-provider-datadog/issues/1651

It seems like the API is lower-casing service names, so when your
terraform code has a service called, for instance, "Foo", it would
accept the name, but in the backend convert it to "foo". That causes
problems with terraform, which stored foo locally, and then it tries to
recreate the entire resource over and over:

```diff
-/+ resource "datadog_service_definition_yaml" "foo" {
      ~ id                 = "foo" -> (known after apply)
      ~ service_definition = <<-EOT # forces replacement
          - dd-service: foo
          + dd-service: Foo
            integrations:
                pagerduty: https://XXX.pagerduty.com/service-directory/XXX
            schema-version: v2
            tags:
                - managed-by:terraform
            team: data-platform
        EOT
    }
```

This commit introduces a change to verify if a resource should be
targeted to "replacement" only if the old and new service names are
different in content.

Additionally, it also stores lowercased dd-service values on terraform
state so that it guarantees this won't be showing in all diffs for every
single terraform plan since the file definitions will be consistently
different than the remote state, so it prevents this:

```diff
  ~ resource "datadog_service_definition_yaml" "foo" {
        id                 = "foo"
      ~ service_definition = <<-EOT
          - dd-service: foo
          + dd-service: Foo
            integrations:
                pagerduty: https://XXX.pagerduty.com/service-directory/XXX
            schema-version: v2
            tags:
                - managed-by:terraform
            team: data-platform
        EOT
    }
```