[SEC-5854] Fix acceptance tests for default rules
The default rule acceptance tests weren't actually testing much
previously. The tests did apply the terraform, but did not make any
assertions/checks on them.

These checks would have failed because the existing tests are importing
an infrastructure configuration rule and attempt to set the "decrease
criticality" flag. This, however, is only supported for log detection
rules.

This PR changes the acceptance tests to make sure we run the
configuration on rule of a supported type and adds a check. It also
fixes a bug where the "decrease criticality" flag was returned in the
terraform response for rule types for which the flag isn't supported.