[datadog_user] Re-order role updates (#2346)
This PR proposes that we re-order role updates for datadog_user, so additions happen before removals.

I was testing `datadog_user` and ran into a funny case. If you're modifying yourself, for example to change from
```
[ datadog_role.old_admin_role.id ]
```
to
```
[ datadog_role.new_admin_role.id ]
```
... then if we do removals first, then additions, we'll get locked out and won't be able to finish the operation. It seems safer to first add all the new roles, then remove any old roles.

Note: it's still possible to delete your own role and lock yourself out, but only if the target list of roles doesn't contain the user_access_manage permission.